+++ admin/settings/templates/class-settings-inputs.php
@@ -179,8 +179,10 @@ class Settings_Inputs {
         return 'Super Admin';
       } elseif ( 'manage_options' === $option ) {
         return 'Administrator';
+      } elseif ( 'state_options' === $option ) {
+        return 'Editor/Manager';
       } elseif ( 'edit_private_pages' === $option ) {
-        return 'Editor';
+          return 'Editor';
       } elseif ( 'publish_posts' === $option ) {
         return 'Author';
       } elseif ( 'edit_posts' === $option ) {